Upgrade create-universal-package@^3.2.x
