+++ src/__tests__/notfound.node.js
@@ -23,25 +23,26 @@
 import test from 'tape-cup';
 import React from 'react';
 import {renderToString as render} from 'react-dom/server';
-import {Router, Route} from '../../server';
+import {Router, Route, NotFound} from "../server";
 
-test('matches as expected', t => {
-  const Hello = () => <div>Hello</div>;
+test('sets code', t => {
+  const Hello = () => (
+    <NotFound>
+      <div>Hello</div>
+    </NotFound>
+  );
+  const state = {code: 0};
+  const ctx = {
+    setCode(code) {
+      state.code = code;
+    },
+  };
   const el = (
-    <Router location="/">
-      <Route path="/" component={Hello} />
+    <Router location="/" context={ctx}>
+      <Route component={Hello} />
     </Router>
   );
-  t.ok(/Hello/.test(render(el)), 'renders matched route in server');
-  t.end();
-});
-test('misses as expected', t => {
-  const Hello = () => <div>Hello</div>;
-  const el = (
-    <Router location="/foo">
-      <Route path="/bar" component={Hello} />
-    </Router>
-  );
-  t.ok(!/Hello/.test(render(el)), 'does not render unmatched route');
+  t.ok(/Hello/.test(render(el)), 'matches');
+  t.equals(state.code, 404, 'sets code');
   t.end();
 });
