+++ src/__tests__/route.browser.js
@@ -24,22 +24,29 @@
 import test from 'tape-cup';
 import React from 'react';
 import ReactDOM from 'react-dom';
-import {Router, Switch, Route} from '../../browser';
+import {Router, Route} from "../browser";
 
 test('matches as expected', t => {
   const root = document.createElement('div');
   const Hello = () => <div>Hello</div>;
-  const Hi = () => <div>Hi</div>;
   const el = (
     <Router>
-      <Switch>
-        <Route path="/" component={Hello} />
-        <Route path="/" component={Hi} />
-      </Switch>
+      <Route path="/" component={Hello} />
     </Router>
   );
   ReactDOM.render(el, root);
-  t.ok(/Hello/.test(root.innerHTML), 'matches first');
-  t.ok(!/Hi/.test(root.innerHTML), 'does not match second');
+  t.ok(/Hello/.test(root.innerHTML), 'renders matched route');
+  t.end();
+});
+test('misses as expected', t => {
+  const root = document.createElement('div');
+  const Hello = () => <div>Hello</div>;
+  const el = (
+    <Router>
+      <Route exact path="/bar" component={Hello} />
+    </Router>
+  );
+  ReactDOM.render(el, root);
+  t.ok(!/Hello/.test(root.innerHTML), 'does not render unmatched route');
   t.end();
 });
