+++ src/__tests__/route.node.js
@@ -23,22 +23,25 @@
 import test from 'tape-cup';
 import React from 'react';
 import {renderToString as render} from 'react-dom/server';
-import {Router, Route, Redirect} from '../../server';
+import {Router, Route} from "../server";
 
-test('redirects to a new URL', t => {
+test('matches as expected', t => {
   const Hello = () => <div>Hello</div>;
-  const Moved = () => <Redirect to="/hello" />;
-  const state = {code: 0};
-  const ctx = state;
   const el = (
-    <Router location="/" context={ctx}>
-      <div>
-        <Route path="/" component={Moved} />
-        <Route path="/hello" component={Hello} />
-      </div>
+    <Router location="/">
+      <Route path="/" component={Hello} />
     </Router>
   );
-  render(el);
-  t.equals(state.code, 307, 'sets code');
+  t.ok(/Hello/.test(render(el)), 'renders matched route in server');
+  t.end();
+});
+test('misses as expected', t => {
+  const Hello = () => <div>Hello</div>;
+  const el = (
+    <Router location="/foo">
+      <Route path="/bar" component={Hello} />
+    </Router>
+  );
+  t.ok(!/Hello/.test(render(el)), 'does not render unmatched route');
   t.end();
 });
